Fixed passing root size when selecting a disk and rendering of elements that take up a small amount of space relative to the total disk size.

+++ Source/Filesystem.cc
@@ -52,7 +52,7 @@ namespace Filesystem {
 		return logicalDrives;
 	}
 
-	std::pair<ULONGLONG, ULONGLONG> GetDriveSpace(std::string_view driveLetter) {
+	std::pair<size_t, size_t> GetDriveSpace(std::string_view driveLetter) {
 		ULARGE_INTEGER bytesTotal, bytesFree;
 		::GetDiskFreeSpaceEx(driveLetter.data(), nullptr, &bytesTotal, &bytesFree);
