+++ Source/Filesystem.h
@@ -10,7 +10,7 @@ namespace Filesystem {
 	};
 
 	std::vector<std::string> GetLogicalDrives();
-	std::pair<ULONGLONG, ULONGLONG> GetDriveSpace(std::string_view driveLetter);
+	std::pair<size_t, size_t> GetDriveSpace(std::string_view driveLetter);
 
 	Tree::Node<Entry> BuildTree(const std::filesystem::path& path, std::atomic<size_t>& progress);
 	Tree::Node<Entry> ParallelBuildTree(const std::filesystem::path& path, std::atomic<size_t>& progress);
